Fix integer overflow on ARMv5
When compiling against ARMv5 (`go run make.go --arch=arm --arm=5`) the `1 << 32` overflows:

    tmp/build-gopath-nosqlite-armv5/src/camlistore.org/pkg/blobserver/b2/b2.go:201: constant 4294967296 overflows int

Cast both sides to uint64 during the comparison to avoid this

+++ pkg/blobserver/b2/b2.go
@@ -198,7 +198,7 @@ func (s *Storage) Fetch(br blob.Ref) (rc io.ReadCloser, size uint32, err error) 
 		return nil, 0, errors.New("b2: remote ContentSHA1 mismatch")
 	}
 
-	if fi.ContentLength >= 1<<32 {
+	if int64(fi.ContentLength) >= int64(1<<32) {
 		r.Close()
 		return nil, 0, errors.New("object larger than a uint32")
 	}
